Download Multiple DMRs

Returns DMRs in a nested list using echoGetEffluent(). Uses a dataframe with a column of p_id numbers. Please note that p_id's are case sensitive.

Usage

downloadDMRs(df, idColumn, pBar = TRUE, verbose = FALSE, ...)

Arguments

df

dataframe with column of id numbers

idColumn

unquoted string, name of column containing the p_id permit numbers

pBar

logical, display a progress bar? Defaults to TRUE

verbose

logical, indicating whether to provide processing and retrieval messages. Defaults to FALSE. Suggest leaving this FALSE if pBar = TRUE.

...

additional arguments passed to echoGetEffluent

Value

dataframe df, with a column containing the discharge monitoring reports downloaded with echoGetEffluentSummary

Examples

# \donttest{
## This example requires an internet connection to run

## Retrieve multiple DMRs for flow

df <- tibble::tibble("id" = c('TX0119407', 'TX0132187'))
df <- downloadDMRs(df, id)
#> Warning: `progress_estimated()` was deprecated in dplyr 1.0.0.
#> ℹ The deprecated feature was likely used in the echor package.
#>   Please report the issue at <https://github.com/mps9506/echor/issues>.
#> Request failed [500]. Retrying in 1 seconds...
#> Request failed [503]. Retrying in 2.1 seconds...
#> Error in mutate(., dmr = purrr::pmap(data, ~{    pb$tick()$print()    Sys.sleep(5)    echoGetEffluent(p_id = .x, verbose = verbose, dots_user)})): ℹ In argument: `dmr = purrr::pmap(...)`.
#> Caused by error in `purrr::pmap()`:
#> ℹ In index: 1.
#> Caused by error in `resp_check()`:
#> ! Service Unavailable (HTTP 503).
# }