Downloads EPA ECHO DMR records of dischargers with NPDES permits

Uses EPA ECHO API to download the Discharge Monitoring Record (DMR) of a single plant, identified with p_id. Please note that the p_id is case sensitive.

echoGetEffluent(p_id, verbose = FALSE, ...)

Arguments

p_id
Character string specify the identifier for the service. Required. Case sensitive.

verbose
Logical, indicating whether to provide processing and retrieval messages. Defaults to FALSE

...
Further arguments passed on as query parameters sent to EPA's ECHO API. For more options see: https://echo.epa.gov/tools/web-services/effluent-charts#!/Effluent_Charts/get_eff_rest_services_get_effluent_chart

Value

Returns a dataframe.

Examples

## This example requires an internet connection to run

## Retrieve single DMR for flow

echoGetEffluent(p_id = 'tx0119407', parameter_code = '50050')

#> # A tibble: 0 x 61
#> # ... with 61 variables: activity_id <chr>, npdes_id <chr>, version_nmbr <chr>,
#> #   perm_feature_id <chr>, perm_feature_nmbr <chr>,
#> #   perm_feature_type_code <chr>, perm_feature_type_desc <chr>,
#> #   limit_set_id <chr>, limit_set_schedule_id <chr>, limit_id <chr>,
#> #   limit_type_code <chr>, limit_begin_date <date>, limit_end_date <date>,
#> #   nmbr_of_submission <chr>, parameter_code <chr>, parameter_desc <chr>,
#> #   monitoring_location_code <chr>, monitoring_location_desc <chr>,
#> #   stay_type_code <chr>, stay_type_desc <chr>, limit_value_id <chr>,
#> #   limit_value_type_code <chr>, limit_value_type_desc <chr>,
#> #   limit_value_nmbr <dbl>, limit_unit_code <chr>, limit_unit_desc <chr>,
#> #   standard_unit_code <chr>, standard_unit_desc <chr>,
#> #   limit_value_standard_units <dbl>, statistical_base_code <chr>,
#> #   statistical_base_short_desc <chr>, statistical_base_type_code <chr>,
#> #   statistical_base_type_desc <chr>, limit_value_qualifier_code <chr>,
#> #   stay_value_nmbr <chr>, dmr_event_id <chr>,
#> #   monitoring_period_end_date <date>, dmr_form_value_id <chr>,
#> #   value_type_code <chr>, value_type_desc <chr>, dmr_value_id <chr>,
#> #   dmr_value_nmbr <dbl>, dmr_unit_code <chr>, dmr_unit_desc <chr>,
#> #   dmr_value_standard_units <dbl>, dmr_value_qualifier_code <chr>,
#> #   value_received_date <date>, days_late <chr>, nodi_code <chr>,
#> #   nodi_desc <chr>, exceedence_pct <dbl>, npdes_violation_id <chr>,
#> #   violation_code <chr>, violation_desc <chr>, rnc_detection_code <chr>,
#> #   rnc_detection_desc <chr>, rnc_detection_date <chr>,
#> #   rnc_resolution_code <chr>, rnc_resolution_desc <chr>,
#> #   rnc_resolution_date <chr>, violation_severity <chr>