Downloads EPA ECHO permitted air emitter information

Returns a dataframe or simplefeature dataframe of permitted facilities returned by the query. Uses EPA's ECHO API: https://echo.epa.gov/tools/web-services/facility-search-air#!/Facilities/get_air_rest_services_get_facility_info

echoAirGetFacilityInfo(output = "df", verbose = FALSE, ...)

Arguments

output
Character string specifying output format. output = 'df' for a dataframe or output = 'sf' for a simple features spatial dataframe. See (https://CRAN.R-project.org/package=sf) for more information about simple features.

verbose
Logical, indicating whether to provide processing and retrieval messages. Defaults to FALSE

...
Further arguments passed as query parameters in request sent to EPA ECHO's API. For more options see: https://echo.epa.gov/tools/web-services/facility-search-water#!/Facility_Information/get_air_rest_services_get_facility_info for a complete list of parameter options. Examples provided below.

Value

dataframe or sf dataframe suitable for plotting

Examples

## These examples require an internet connection to run

## Retrieve table of facilities by bounding box
echoAirGetFacilityInfo(xmin = '-96.407563',
ymin = '30.554395',
xmax = '-96.25947',
ymax = '30.751984',
output = 'df')

#> # A tibble: 40 x 29
#>    ObjectId AIRName SourceID AIRStreet AIRCity AIRState LocalControlReg~ AIRZip
#>    <chr>    <chr>   <chr>    <chr>     <chr>   <chr>    <lgl>            <chr> 
#>  1 1        AGGIE ~ 0600000~ 111 COLL~ COLLEG~ TX       NA               77840~
#>  2 2        ALENCO~ TX00000~ 615 CARS~ BRYAN   TX       NA               77801 
#>  3 3        ALL SE~ 0600000~ 2501 TEX~ COLLEG~ TX       NA               77840 
#>  4 4        BLUEBO~ TX00000~ HWY. 60,~ COLLEG~ TX       NA               77840 
#>  5 5        BRIARC~ 0600000~ 1887 BRI~ BRYAN   TX       NA               77802 
#>  6 6        BRYAN ~ TX00000~ 1500 IND~ BRYAN   TX       NA               77803 
#>  7 7        BRYAN ~ 0600000~ 1803 HOL~ COLLEG~ TX       NA               77840 
#>  8 8        BRYAN ~ TX00000~ 3747 OLD~ BRYAN   TX       NA               77808~
#>  9 9        CITGO ~ TX00000~ 1714 FIN~ BRYAN   TX       NA               77801 
#> 10 10       CITY O~ TX00000~ 1.5 MI W~ BRYAN   TX       NA               77801 
#> # ... with 30 more rows, and 21 more variables: RegistryID <chr>,
#> #   AIRCounty <chr>, AIREPARegion <chr>, FacFederalAgencyCode <lgl>,
#> #   FacFIPSCode <chr>, FacLat <chr>, FacLong <chr>, AIRPrograms <chr>,
#> #   AIRMacts <chr>, AIRStatus <chr>, ChesapeakeBayFlag <lgl>, TRIIDs <chr>,
#> #   FacNaaFlag <lgl>, FacTRIAIRReleases <chr>, FacEisAIRReleases <chr>,
#> #   FacGHGCO2Releases <lgl>, AIRIeaCnt <chr>, AIRLastIeaDate <chr>,
#> #   PctilePmUs <chr>, FacMapIcon <chr>, Map <lgl>

## Retrieve a simple features dataframe by bounding box
spatialdata <- echoAirGetFacilityInfo(xmin = '-96.407563',
ymin = '30.554395',
xmax = '-96.25947',
ymax = '30.751984',
output = 'sf')

#> [1] "# Status message: Success"           "# Status message: OK"               
#> [3] "# Status message: Success: (200) OK"

#> No encoding supplied: defaulting to UTF-8.