Downloads EPA ECHO DMR records of dischargers with NPDES permits

Uses EPA ECHO API to download the Discharge Monitoring Record (DMR) of a single plant, identified with p_id. Please note that the p_id is case sensitive.

echoGetEffluent(p_id, verbose = FALSE, ...)

Arguments

p_id
Character string specify the identifier for the service. Required. Case sensitive.

verbose
Logical, indicating whether to provide processing and retrieval messages. Defaults to FALSE

...
Further arguments passed on as query parameters sent to EPA's ECHO API. For more options see: https://echo.epa.gov/tools/web-services/effluent-charts#!/Effluent_Charts/get_eff_rest_services_get_effluent_chart

Value

Returns a dataframe.

Examples

# \donttest{
## This example requires an internet connection to run

## Retrieve single DMR for flow

echoGetEffluent(p_id = 'tx0119407', parameter_code = '50050')

#> # A tibble: 80 × 61
#>    activity_id npdes_id  version_nmbr perm_feature_id perm_feature_nmbr
#>    <chr>       <chr>     <chr>        <chr>           <chr>            
#>  1 3600178396  TX0119407 4            3600049681      001              
#>  2 3600178396  TX0119407 4            3600049681      001              
#>  3 3600178396  TX0119407 4            3600049681      001              
#>  4 3600178396  TX0119407 4            3600049681      001              
#>  5 3600178396  TX0119407 4            3600049681      001              
#>  6 3600178396  TX0119407 4            3600049681      001              
#>  7 3600178396  TX0119407 4            3600049681      001              
#>  8 3600178396  TX0119407 4            3600049681      001              
#>  9 3600178396  TX0119407 4            3600049681      001              
#> 10 3600178396  TX0119407 4            3600049681      001              
#> # … with 70 more rows, and 56 more variables: perm_feature_type_code <chr>,
#> #   perm_feature_type_desc <chr>, limit_set_id <chr>,
#> #   limit_set_schedule_id <chr>, limit_id <chr>, limit_type_code <chr>,
#> #   limit_begin_date <chr>, limit_end_date <chr>, nmbr_of_submission <chr>,
#> #   parameter_code <chr>, parameter_desc <chr>, monitoring_location_code <chr>,
#> #   monitoring_location_desc <chr>, stay_type_code <chr>, stay_type_desc <chr>,
#> #   limit_value_id <chr>, limit_value_type_code <chr>, …

# }