Downloads public water system information

Returns a dataframe of permitted public water systems returned by the query. Uses EPA's ECHO API: https://echo.epa.gov/tools/web-services/facility-search-drinking-water#!/Safe_Drinking_Water/get_sdw_rest_services_get_systems.

echoSDWGetSystems(verbose = FALSE, ...)

Arguments

verbose
Logical, indicating whether to provide processing and retrieval messages. Defaults to FALSE

...
Further arguments passed as query parameters in request sent to EPA ECHO's API. For more options see: https://echo.epa.gov/tools/web-services/facility-search-drinking-water#!/Safe_Drinking_Water/get_sdw_rest_services_get_systems for a complete list of parameter options. Examples provided below.

Value

returns a dataframe

Examples

# \donttest{
## These examples require an internet connection to run
echoSDWGetSystems(p_co = "Brazos", p_st = "tx")

#> # A tibble: 39 x 66
#>    PWSName PWSId CitiesServed StateCode ZipCodesServed CountiesServed EPARegion
#>    <chr>   <chr> <chr>        <chr>     <chr>          <chr>          <chr>    
#>  1 ABBATE~ TX02~ ""           TX        ""             Brazos         06       
#>  2 AL LEO~ TX02~ ""           TX        ""             Brazos         06       
#>  3 B J'S ~ TX02~ ""           TX        ""             Brazos         06       
#>  4 BENCHL~ TX02~ ""           TX        ""             Brazos         06       
#>  5 BOBBIT~ TX02~ ""           TX        ""             Brazos         06       
#>  6 BRUSHY~ TX02~ ""           TX        ""             Brazos         06       
#>  7 CAFE S~ TX02~ ""           TX        ""             Brazos         06       
#>  8 CAROUS~ TX02~ ""           TX        ""             Brazos         06       
#>  9 CITY O~ TX02~ ""           TX        ""             Brazos         06       
#> 10 CITY O~ TX02~ ""           TX        ""             Brazos         06       
#> # ... with 29 more rows, and 59 more variables: IndianCountry <chr>,
#> #   PWSTypeCode <chr>, PWSTypeDesc <chr>, PrimarySourceCode <chr>,
#> #   PrimarySourceDesc <chr>, PopulationServedCount <dbl>,
#> #   PWSActivityCode <chr>, PWSActivityDesc <chr>, OwnerTypeCode <chr>,
#> #   OwnerDesc <chr>, QtrsWithVio <dbl>, QtrsWithSNC <dbl>,
#> #   SeriousViolator <chr>, HealthFlag <chr>, MrFlag <chr>, PnFlag <chr>,
#> #   OtherFlag <chr>, NewVioFlg <chr>, RulesVio3yr <dbl>, RulesVio <dbl>,
#> #   Viopaccr <dbl>, Vioremain <dbl>, Viofeanot <dbl>, Viortcfea <dbl>,
#> #   Viortcnofea <dbl>, Ifea <dbl>, Feas <dbl>, SDWDateLastIea <date>,
#> #   SDWDateLastIeaEPA <date>, SDWDateLastIeaSt <date>, SDWDateLastFea <date>,
#> #   SDWDateLastFeaEPA <date>, SDWDateLastFeaSt <date>,
#> #   SDWAContaminantsInViol3yr <chr>, SDWAContaminantsInCurViol <chr>,
#> #   PbAle <dbl>, CuAle <dbl>, Rc350Viol <dbl>, DfrUrl <chr>, FIPSCodes <chr>,
#> #   SNC <chr>, GwSwCode <chr>, SDWA3yrComplQtrsHistory <chr>,
#> #   SDWAContaminants <chr>, PbViol <dbl>, CuViol <dbl>,
#> #   LeadAndCopperViol <dbl>, TRIbalFlag <dbl>, FeaFlag <dbl>, IeaFlag <dbl>,
#> #   SNCFlag <dbl>, CurrVioFlag <dbl>, VioFlag <dbl>, Insp5yrFlag <dbl>,
#> #   ServiceAreaTypeCode <chr>, ServiceAreaTypeDesc <chr>,
#> #   ViolationCategories <chr>, Sansurvey5yr <dbl>, DateLastSansurvey <date>

# }