Downloads EPA ECHO water facility information

Returns a dataframe or simplefeature dataframe of permitted facilities returned by the query. Uses EPA's ECHO API: https://echo.epa.gov/tools/web-services/facility-search-water.

echoWaterGetFacilityInfo(output = "df", verbose = FALSE, ...)

Arguments

output
Character string specifying output format. output = 'df' for a dataframe or output = 'sf' for a simple features spatial dataframe. See (https://CRAN.R-project.org/package=sf) for more information about simple features.

verbose
Logical, indicating whether to provide processing and retrieval messages. Defaults to FALSE

…
Further arguments passed as query parameters in request sent to EPA ECHO's API. For more options see: https://echo.epa.gov/tools/web-services/facility-search-water#!/Facility_Information/get_cwa_rest_services_get_facility_info for a complete list of parameter options. Examples provided below.

Value

returns a dataframe or simple features dataframe

Examples

## These examples require an internet connection to run

## Retrieve table of facilities by bounding box
echoWaterGetFacilityInfo(xmin = '-96.407563',
ymin = '30.554395',
xmax = '-96.25947',
ymax = '30.751984',
output = 'df')

#> # A tibble: 60 x 28
#>    ObjectId CWPName SourceID CWPStreet CWPCity CWPState CWPStateDistrict CWPZip
#>    <chr>    <chr>   <chr>    <chr>     <chr>   <chr>    <lgl>            <chr> 
#>  1 1        AGGIE ~ TX01321~ 800 FT S~ COLLEG~ TX       NA               77845 
#>  2 2        AGRIVE~ TX01212~ SWISHER ~ BRYAN   TX       NA               00000 
#>  3 3        ATKINS~ TX00279~ 601 ATKI~ BRYAN   TX       NA               77801 
#>  4 4        ATOFIN~ TX01088~ SW OF TH~ BRYAN   TX       NA               77801 
#>  5 5        BARTLE~ TX01204~ SWISHER ~ AMARIL~ TX       NA               00000 
#>  6 6        BOSSIE~ LAG8301~ 3228 BAR~ BENTON  LA       NA               71111 
#>  7 7        BRAYTO~ TX00908~ 1595 NUC~ COLLEG~ TX       NA               77843~
#>  8 8        BRAZOS~ TXR0401~ URBANIZE~ BRYAN   TX       NA               77803 
#>  9 9        BRM 5   TXG1119~ 1391 OLD~ BRYAN-~ TX       NA               77845 
#> 10 10       BROADS~ TXR10F5~ NW OF AT~ BRYAN   TX       NA               77807 
#> # ... with 50 more rows, and 20 more variables: MasterExternalPermitNmbr <chr>,
#> #   RegistryID <chr>, CWPCounty <chr>, CWPEPARegion <chr>, FacDerivedHuc <chr>,
#> #   FacLat <chr>, FacLong <chr>, CWPTotalDesignFlowNmbr <chr>,
#> #   CWPActualAverageFlowNmbr <chr>, ReceivingMs4Name <lgl>,
#> #   AssociatedPollutant <lgl>, MsgpPermitType <lgl>, CWPPermitStatusDesc <chr>,
#> #   CWPPermitTypeDesc <chr>, CWPIssueDate <chr>, CWPEffectiveDate <chr>,
#> #   CWPExpirationDate <chr>, CWPSNCStatusDate <chr>,
#> #   CWPStateWaterBodyCode <chr>, FacMapFlg <chr>

## Retrieve a simple features dataframe by bounding box
spatialdata <- echoWaterGetFacilityInfo(xmin = '-96.407563',
ymin = '30.554395',
xmax = '-96.25947',
ymax = '30.751984',
output = 'sf')

#> No encoding supplied: defaulting to UTF-8.