Downloads EPA ECHO water facility information

Returns a dataframe or simplefeature dataframe of permitted facilities returned by the query. Uses EPA's ECHO API: https://echo.epa.gov/tools/web-services/facility-search-water.

echoWaterGetFacilityInfo(output = "df", verbose = FALSE, ...)

Arguments

output
Character string specifying output format. output = 'df' for a dataframe or output = 'sf' for a simple features spatial dataframe. See (https://CRAN.R-project.org/package=sf) for more information about simple features.

verbose
Logical, indicating whether to provide processing and retrieval messages. Defaults to FALSE

...
Further arguments passed as query parameters in request sent to EPA ECHO's API. For more options see: https://echo.epa.gov/tools/web-services/facility-search-water#!/Facility_Information/get_cwa_rest_services_get_facility_info for a complete list of parameter options. Examples provided below.

Value

returns a dataframe or simple features dataframe

Examples

# \donttest{
## These examples require an internet connection to run

## Retrieve table of facilities by bounding box
echoWaterGetFacilityInfo(xmin = '-96.407563',
ymin = '30.554395',
xmax = '-96.25947',
ymax = '30.751984',
output = 'df')

#> Warning: One or more parsing issues, see `problems()` for details

#> # A tibble: 419 × 26
#>    CWPName     SourceID  CWPStreet      CWPCity CWPState CWPStateDistrict CWPZip
#>    <chr>       <chr>     <chr>          <chr>   <chr>    <chr>            <chr> 
#>  1 3444 MAHOG… TXR1509EW SOUTHWEST OF … BRYAN   TX       ""               77807 
#>  2 5113 INVER… TXR1542EK 5113 INVERNES… BRYAN   TX       ""               77802 
#>  3 777 GRAHAM… TXR1578AE 777 GRAHAM RD  COLLEG… TX       ""               77845 
#>  4 8.5 ACRE D… TXR1535BW NORTHWEST INT… BRYAN   TX       ""               77803 
#>  5 ACADEMY CO… TXR1508FA 1326 PRAIRIE … BRYAN   TX       ""               77803 
#>  6 ACE TOWNHO… TXR15667I 2136 CHESTNUT… COLLEG… TX       ""               77845…
#>  7 AGGIE PARK… TXR1543FV 800 THROCKMOR… COLLEG… TX       ""               77840 
#>  8 AGRILIFE P… TXR1582BA 950 AGRONOMY … COLLEG… TX       ""               77845 
#>  9 ALENCO WIN… TXR05U808 615 W CARSON … BRYAN   TX       ""               77801…
#> 10 ANDREWS OR… TXR15193R 1098 ARRINGTO… COLLEG… TX       ""               77845 
#> # … with 409 more rows, and 19 more variables: MasterExternalPermitNmbr <chr>,
#> #   RegistryID <chr>, CWPCounty <chr>, CWPEPARegion <chr>, FacDerivedHuc <chr>,
#> #   FacLat <dbl>, FacLong <dbl>, CWPTotalDesignFlowNmbr <dbl>,
#> #   CWPActualAverageFlowNmbr <dbl>, ReceivingMs4Name <chr>,
#> #   AssociatedPollutant <chr>, MsgpPermitType <chr>, CWPPermitStatusDesc <chr>,
#> #   CWPPermitTypeDesc <chr>, CWPIssueDate <date>, CWPEffectiveDate <date>,
#> #   CWPExpirationDate <date>, CWPSNCStatusDate <date>, …

## Retrieve a simple features dataframe by bounding box
spatialdata <- echoWaterGetFacilityInfo(xmin = '-96.407563',
ymin = '30.554395',
xmax = '-96.25947',
ymax = '30.751984',
output = 'sf')

# }