Downloads EPA ECHO water facility information

Returns a dataframe or simplefeature dataframe of permitted facilities returned by the query. Uses EPA's ECHO API: https://echo.epa.gov/tools/web-services/facility-search-water.

Usage

echoWaterGetFacilityInfo(output = "df", verbose = FALSE, ...)

Arguments

output

Character string specifying output format. output = 'df' for a dataframe or output = 'sf' for a simple features spatial dataframe. See (https://CRAN.R-project.org/package=sf) for more information about simple features.

verbose

Logical, indicating whether to provide processing and retrieval messages. Defaults to FALSE

...

Further arguments passed as query parameters in request sent to EPA ECHO's API. For more options see: https://echo.epa.gov/tools/web-services/facility-search-water#!/Facility_Information/get_cwa_rest_services_get_facility_info for a complete list of parameter options. Examples provided below.

Value

returns a dataframe or simple features dataframe

Examples

# \donttest{
## These examples require an internet connection to run

## Retrieve table of facilities by bounding box
echoWaterGetFacilityInfo(xmin = '-96.407563',
ymin = '30.554395',
xmax = '-96.25947',
ymax = '30.751984',
output = 'df')
#> # A tibble: 603 × 26
#>    CWPName           SourceID CWPStreet CWPCity CWPState CWPStateDistrict CWPZip
#>    <chr>             <chr>    <chr>     <chr>   <chr>    <chr>            <chr> 
#>  1 066-3344 BORLAUG… TXR1579… 600 JOHN… COLLEG… TX       NA               77843…
#>  2 2900 SPECTOR DRI… TXR1571… 2900 SPE… BRYAN   TX       NA               77802 
#>  3 2901 SPECTOR DRI… TXR1531… 2901 SPE… BRYAN   TX       NA               77802 
#>  4 3444 MAHOGANY DR… TXR1509… SOUTHWES… BRYAN   TX       NA               77807 
#>  5 5113 INVERNESS DR TXR1542… 5113 INV… BRYAN   TX       NA               77802 
#>  6 777 GRAHAM RD     TXR1578… 777 GRAH… COLLEG… TX       NA               77845 
#>  7 8.5 ACRE DEVELOP… TXR1535… NORTHWES… BRYAN   TX       NA               77803 
#>  8 ACADEMY CONDOMIN… TXR1508… 1326 PRA… BRYAN   TX       NA               77803 
#>  9 ACADEMY CONDOS    TXR1512… 1326 PRA… BRYAN   TX       NA               77803 
#> 10 ACADEMY PRESERVE… TXR1504… SEC OF O… BRYAN   TX       NA               77803 
#> # ℹ 593 more rows
#> # ℹ 19 more variables: MasterExternalPermitNmbr <chr>, RegistryID <chr>,
#> #   CWPCounty <chr>, CWPEPARegion <chr>, FacDerivedHuc <chr>,
#> #   CWPNAICSCodes <chr>, FacLat <dbl>, FacLong <dbl>,
#> #   CWPTotalDesignFlowNmbr <dbl>, DschToMs4 <chr>, ExposedActivity <chr>,
#> #   NPDESDataGroupsDescs <chr>, MsgpFacilityInspctnSmmry <chr>,
#> #   MsgpCorrectiveActionSmmry <chr>, AIRIDs <chr>, NPDESIDs <chr>, …

## Retrieve a simple features dataframe by bounding box
spatialdata <- echoWaterGetFacilityInfo(xmin = '-96.407563',
ymin = '30.554395',
xmax = '-96.25947',
ymax = '30.751984',
output = 'sf')

# }